Consolidating new.go and existing.go into creation.go
